Refactor projects route
