+++ apps/frontend/src/pages/Projects.jsx
@@ -1,24 +1,34 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { Card } from '../components/Card';
 import { Link } from 'react-router-dom';
+import ProjectDataService from '../services/projects';
+
 import { StyledContainer, StyledH1, StyledH2 } from 'ui';
 
-export default function Projects({ ...props }) {
-  let projects = props.projects;
+export default function Projects() {
+  const [projects, setProjects] = useState([]);
+
+  useEffect(() => {
+    retrieveProjects();
+  }, []);
+
+  const retrieveProjects = () => {
+    ProjectDataService.getAll()
+      .then((response) => {
+        setProjects(response.data.data);
+      })
+      .catch((e) => console.log(e));
+  };
   return (
-    <ul>
-      {projects.map(({ id, imageUrl, projectName, description }) => (
-        <li key={{ id }}>
-          <Link to={id}>
-            <Card
-              id={id}
-              imageUrl={imageUrl}
-              projectName={projectName}
-              description={description}
-            />
-          </Link>
-        </li>
+    <>
+      {projects.map((project) => (
+        <Card
+          id={project._id}
+          imageUrl={project.imageUrl}
+          projectName={project.projectName}
+          description={project.description}
+        />
       ))}
-    </ul>
+    </>
   );
 }
